sgi-gr2: store palette indices in the framebuffer, apply the palette at scanout
The RGB was resolved at draw time, so a pixel took whatever the RAMDAC held
at that moment.  The DDX draws the root before it programs the entries the
root uses, so the stipple came out black (indices 0 and 64 were still
unprogrammed when it ran) even though both entries are programmed a moment
later.

Store the index in the framebuffer, as the hardware does, and expand it
through the palette in sgi_gr2_update_display.  A pixel drawn before its
entry is programmed now shows the entry's final colour.

Measured on the default boot: the stipple renders 87.5% black plus 12.5%
(47,47,47) = index 64's programmed value, exactly the 1-in-8 grain of
pattern 0x10101010.  The note-18 gate stays green (pre black, solid red,
blue, red again).

The scanout-bars test seam keeps its bars by loading the eight bar colours
into palette entries 0..7 and writing those indices.

+++ include/hw/misc/sgi_gr2.h
@@ -162,10 +162,13 @@ struct SGIGr2State {
     bool xmap_ready;
 
     /* Scanout (P0.4 step a): a QEMU display surface proving the output stage
-     * before the RE3 producer.  `scanout` is the framebuffer VC1 would scan
-     * (host xRGB32); VC1/RE3 will read it once the producer is emulated. */
+     * before the RE3 producer.  `scanout` holds palette INDICES, exactly as the
+     * hardware framebuffer does: the RAMDAC is applied when the buffer is
+     * scanned out, not when it is drawn.  That ordering matters — the DDX draws
+     * the root before it programs the entries the root uses, so resolving RGB
+     * at draw time would bake in whatever the palette happened to hold then. */
     QemuConsole *con;
-    uint32_t *scanout;
+    uint8_t *scanout;
     bool scanout_bars; /* fill a colour-bar test pattern (P0.4 step a) */
 
     /* RE3 producer colour latch + RAMDAC palette (8-bit mode).  In the 8-bit
